Mark only the required layer types as dirty
